Fix layout and camera bugs and add fab to main activity

+++ app/app/src/main/java/com/treble/appetite/meal/data/database/room/MealEntity.kt
@@ -8,7 +8,7 @@ import java.util.*
 
 @Entity
 data class MealEntity(
-    @PrimaryKey(autoGenerate = true) var id: Int,
+    @PrimaryKey(autoGenerate = true) var id: Long,
     @Nullable
     var mealDate: Date?,
     var beforeImagePath: String,
